+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+## [3.2.0]
+### Added
+ - New generic component for lists
+### Changed
+ - All lists are now filterable
+
 ## [3.1.0]
 ### Added
  - Fixed box with link to "Ouvidoria"